Skip experiments that has already been done.

+++ tools/benchs.sh
@@ -9,38 +9,53 @@ output_folder=results-$command_file_name
 tmp_file=tmp
 
 if [ -e $output_folder ]; then
-    echo -e "Warning : $output_folder folder already exists."
-    echo -e "Please remove it manually to avoid making mistakes."
-    echo -e "Aborting script."
-    exit 1
+    echo -e "$output_folder folder already exists."
+    echo -e "Benchs will be resumed."
+else
+    mkdir $output_folder
 fi
-
-mkdir $output_folder
 
 for cores in $cores_list; do
     export HERMIT_CPUS=$cores
     export OMP_NUM_THREADS=$cores
     echo -e "Running benchs for $cores cores.\n===="
-    mkdir $output_folder/$cores-cores
+
+    core_folder=$output_folder/$cores-cores
+    if [ -e $core_folder ]; then
+        echo -e "$core_folder already exists."
+        echo -e "Benchs that have already been done will be skipped.\n"
+    else
+        mkdir $core_folder
+    fi
+    
     cat $command_file | while read timeout_args; do
         prog_name=$(echo $timeout_args | awk '{print $4}')
         prog_name=$(basename $prog_name)
-        output_file=$output_folder/$cores-cores/$prog_name.log
-        touch $output_file
-        echo -e "Issuing command : ./timeout_run.sh $timeout_args"
-        echo -e "Redirecting outputs to $output_file."
-        for((i = 0; i < $n; i++)); do
-            echo -e -n "\r\tCurrent iteration : $i"
-            return_code=1
-            while [ $return_code != "0" ]; do
-                ./timeout_run.sh $timeout_args > $tmp_file
-                return_code=$?
+        output_file=$core_folder/$prog_name.log
+        tmp_output_file=$output_file.tmp
+        if [ -e $output_file ]; then
+            echo -e "$output_file already exist. Experience has already been done."
+            echo -e "Skipping execution of : $timeout_args"
+            echo -e "--"
+        else
+            echo -e "Issuing command : ./timeout_run.sh $timeout_args"
+            echo -e "Redirecting outputs to $output_file."
+            for((i = 0; i < $n; i++)); do
+                echo -e -n "\r\tCurrent iteration : $i"
+                return_code=1
+                while [ $return_code != "0" ]; do
+                    ./timeout_run.sh $timeout_args > $tmp_file
+                    return_code=$?
+                done
+                cat $tmp_file >> $tmp_output_file
             done
-            cat $tmp_file >> $output_file
-        done
-        echo -e "\n--"
+            mv $tmp_output_file $output_file
+            echo -e "\n--"
+        fi
     done
     echo "===="
 done
 
-rm $tmp_file
+if [ -e $tmp_file ]; then
+    rm $tmp_file
+fi
